Update references to exports moved from carddb to cardCatalog.
