+++ src/jobs/update_cards.ts
@@ -19,7 +19,8 @@ import { CardDetails, ColorCategory, DefaultElo } from 'datatypes/Card';
 import { ManaSymbol } from 'datatypes/Mana';
 
 import * as cardutil from '../client/utils/cardutil';
-import { CardMetadata, fileToAttribute, reasonableCard } from '../util/carddb';
+import { CardMetadata, fileToAttribute } from '../util/cardCatalog';
+import { reasonableCard } from '../util/carddb';
 import * as util from '../util/util';
 
 interface ScryfallCard {
